reannotator: merge with rerooter
