+++ setup.py
@@ -20,7 +20,8 @@ setup(name='graftm',
                         'biom-format >=2.1.4',
                         'extern >=0.0.4',
                         'taxtastic >=0.5.4',
-                        'tempdir >=0.6'),
+                        'tempdir >=0.6',
+                        'DendroPy >= 4.1.0'),
       setup_requires=['nose>=1.0'],
       test_suite='nose.collector',
       url='http://geronimp.github.io/graftM',
